requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
PyMuPDF==1.25.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
setuptools==75.8.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
